An example of data.frame export from "clim_extract"

A dataset exported from function clim_extract forclimatic diagrams drawing without frosty months display. It contains the abbreviation, coordinates and altitude of different locations, as well as the values of 3 types of climate data.

Format

A data frame with 30 rows and 18 variables:

No

Serial number of the locations.

Location

Abbreviation of the locations.

Lon

Longitude of the locations in decimal digit.

Lat

Latitude of the locations in decimal digit.

Altitude

Altitude of the locations.

Type

Labels of the climate data, including annual average precipitation,annual average minimum temperature,annual average maximum temperature, and annual extreme minimum temperature.

1

Values of specific type of climate data,the names of columns represent monthly values from January to December.

2

Values of specific type of climate data,the names of columns represent monthly values from January to December.

3

Values of specific type of climate data,the names of columns represent monthly values from January to December.

4

Values of specific type of climate data,the names of columns represent monthly values from January to December.

5

Values of specific type of climate data,the names of columns represent monthly values from January to December.

6

Values of specific type of climate data,the names of columns represent monthly values from January to December.

7

Values of specific type of climate data,the names of columns represent monthly values from January to December.

8

Values of specific type of climate data,the names of columns represent monthly values from January to December.

9

Values of specific type of climate data,the names of columns represent monthly values from January to December.

10

Values of specific type of climate data,the names of columns represent monthly values from January to December.

11

Values of specific type of climate data,the names of columns represent monthly values from January to December.

12

Values of specific type of climate data,the names of columns represent monthly values from January to December.

...

Source

NULL

Details

18 columns of this dataset: "No", "Location", "Lon", "Lat", "Altitude" are information of locations, "Type" are the labels of climate data, and the rest of columns with colnames "1-12" are the climate data values of different months.